+++ ev.c
@@ -1405,8 +1405,8 @@ const char* ev_strerror(int err)
 #line 23 "ev.c"
 ////////////////////////////////////////////////////////////////////////////////
 // FILE:    ev/fs.c
-// SIZE:    23343
-// SHA-256: 5495afdec31a2caca86574c3af0c9e81703e22adaba65fd2c621a91195a8a1fa
+// SIZE:    23400
+// SHA-256: d56fb811ea63bd63fb37837b6a4189ec31079dc47eeac83baa5d54315f229f78
 ////////////////////////////////////////////////////////////////////////////////
 #line 1 "ev/fs.c"
 #include <sys/stat.h>
@@ -1783,7 +1783,7 @@ static void _ev_file_on_read(ev_work_t* work)
     ev_file_t* file = req->file;
     ev_read_t* read_req = &req->req.as_read.read_req;
 
-    req->result = ev_file_read_sync(file, read_req->data.bufs,
+    req->result = ev__fs_readv(file->file, read_req->data.bufs,
         read_req->data.nbuf);
 }
 
@@ -2106,15 +2106,16 @@ int ev_file_seek(ev_file_t* file, ev_fs_req_t* req, int whence, ssize_t offset, 
     return 0;
 }
 
-int ev_file_read(ev_file_t* file, ev_fs_req_t* req, ev_buf_t bufs[],
+ssize_t ev_file_read(ev_file_t* file, ev_fs_req_t* req, ev_buf_t bufs[],
     size_t nbuf, ev_file_cb cb)
 {
-    return _ev_file_read_template(file, req, bufs, nbuf, 0, cb, _ev_file_on_read);
-}
+    if (file->base.loop == NULL)
+    {
+        EV_ASSERT(req == NULL && cb == NULL, "file open in synchronous mode.");
+        return ev__fs_readv(file->file, bufs, nbuf);
+    }
 
-ssize_t ev_file_read_sync(ev_file_t* file, ev_buf_t bufs[], size_t nbuf)
-{
-    return ev__fs_readv(file->file, bufs, nbuf);
+    return _ev_file_read_template(file, req, bufs, nbuf, 0, cb, _ev_file_on_read);
 }
 
 int ev_file_pread(ev_file_t* file, ev_fs_req_t* req, ev_buf_t bufs[],